fix(aivideo): bind entire salvage frame range to descriptor
Hostile RED regressions proved wrong-content renders slipped through
three recipe verifiers: freeze_extension only sampled the suffix
(ignoring the pre-transition prefix), region_crop trusted dimensions
alone, and still_frame accepted any decodable image. Each forgery
rendered successfully before the fix.

Independent descriptor-backed comparisons now bind every claimed frame
to the verified source without reusing the production renderer, while
preserving pass_fds identity for ffmpeg/ffprobe subprocesses:

- freeze_extension compares every pre-transition, transition, and
  extension frame against the source hash stream
- region_crop independently re-renders the declared crop with the same
  codec settings and compares every frame hash
- still_frame decodes the declared timestamp and the produced PNG in a
  common rgb24 space and compares pixels
- background_only keeps the existing FFV1-backed direct comparison

Verification helpers moved into kinocut/aivideo/salvage_checks.py to
keep salvage.py under the 800-line module limit; all functions remain
under 80 lines. Public API is unchanged.
